Update decision-tree.md adding links (#36)
* Update decision-tree.md adding links

Adding links and converting HTML to Markdown for ease of maintenance.

* Update decision-tree.md adding links in page

primary, secondary ownership and contributors

* Update decision-tree.md correcting path

removing /WAI

* Update decision-tree.md spacing & emphasis

formatting

* Update tasks.md - getting rid of main role

This won't show up in the tasks list as it is a duplicate of primary.

* Update visual-designer.md

changes on "Path to Publishing"

* Update content-author.md

changes on "Path to Publishing"

* Update front-end.md (#37)

changes on "Path to Publishing"

* Update content-author.md

Tasks to get started

* Update front-end.md

Tasks to get started

* Update visual-designer.md

Tasks to get started.

* Update content/content-author.md

* Update content/content-author.md

* Update content/content-author.md

* Update content/front-end.md

* Update content/front-end.md

* Update content/visual-designer.md

* Update content/content-author.md

* Update content/visual-designer.md

* Update content/front-end.md

---------

Co-authored-by: Shawn Lawton Henry <shawn@w3.org>